Refactor mock server and client test.
