+++ api/client/client_test.go
@@ -19,6 +19,7 @@ package client
 import (
 	"context"
 	"flag"
+	"fmt"
 	"net"
 	"os"
 	"testing"
@@ -26,95 +27,87 @@ import (
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/gravitational/trace"
+	"github.com/gravitational/trace/trail"
 	"github.com/sirupsen/logrus"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials/insecure"
 
 	"github.com/gravitational/teleport/api/client/proto"
 	"github.com/gravitational/teleport/api/defaults"
 	"github.com/gravitational/teleport/api/types"
 )
 
+func TestMain(m *testing.M) {
+	flag.Parse()
+	if testing.Verbose() {
+		logrus.SetLevel(logrus.DebugLevel)
+	}
+	os.Exit(m.Run())
+}
+
+type pingService struct {
+	*proto.UnimplementedAuthServiceServer
+}
+
+func (s *pingService) Ping(ctx context.Context, req *proto.PingRequest) (*proto.PingResponse, error) {
+	return &proto.PingResponse{}, nil
+}
+
 func TestNew(t *testing.T) {
 	t.Parallel()
 	ctx := context.Background()
-	srv := startMockServer(t)
+	srv := startMockServer(t, &pingService{})
 
 	tests := []struct {
-		desc      string
-		config    Config
-		assertErr require.ErrorAssertionFunc
+		desc         string
+		modifyConfig func(*Config)
+		assertErr    require.ErrorAssertionFunc
 	}{{
-		desc: "successfully dial tcp address.",
-		config: Config{
-			Addrs: []string{srv.Addr()},
-			Credentials: []Credentials{
-				&mockInsecureTLSCredentials{}, // TODO(Joerger) replace insecure credentials
-			},
-			DialOpts: []grpc.DialOption{
-				grpc.WithTransportCredentials(insecure.NewCredentials()), // TODO(Joerger) remove insecure dial option
-			},
-		},
-		assertErr: require.NoError,
+		desc:         "successfully dial tcp address.",
+		modifyConfig: func(c *Config) { /* noop */ },
+		assertErr:    require.NoError,
 	}, {
 		desc: "synchronously dial addr/cred pairs and succeed with the 1 good pair.",
-		config: Config{
-			Addrs: []string{"bad addr", srv.Addr(), "bad addr"},
-			Credentials: []Credentials{
-				&tlsConfigCreds{nil},
-				&mockInsecureTLSCredentials{}, // TODO(Joerger) replace insecure credentials
-				&tlsConfigCreds{nil},
-			},
-			DialOpts: []grpc.DialOption{
-				grpc.WithTransportCredentials(insecure.NewCredentials()), // TODO(Joerger) remove insecure dial option
-			},
+		modifyConfig: func(c *Config) {
+			c.Addrs = append(c.Addrs, "bad addr", "bad addr")
+			c.Credentials = append([]Credentials{&tlsConfigCreds{nil}, &tlsConfigCreds{nil}}, c.Credentials...)
 		},
 		assertErr: require.NoError,
 	}, {
 		desc: "fail to dial with a bad address.",
-		config: Config{
-			DialTimeout: time.Second,
-			Addrs:       []string{"bad addr"},
-			Credentials: []Credentials{
-				&mockInsecureTLSCredentials{}, // TODO(Joerger) replace insecure credentials
-			},
-			DialOpts: []grpc.DialOption{
-				grpc.WithTransportCredentials(insecure.NewCredentials()), // TODO(Joerger) remove insecure dial option
-			},
+		modifyConfig: func(c *Config) {
+			c.Addrs = []string{"bad addr"}
 		},
 		assertErr: func(t require.TestingT, err error, _ ...interface{}) {
 			require.Error(t, err)
-			require.Contains(t, err.Error(), "all connection methods failed")
+			require.ErrorContains(t, err, "all connection methods failed")
 		},
 	}, {
 		desc: "fail to dial with no address or dialer.",
-		config: Config{
-			DialTimeout: time.Second,
-			Credentials: []Credentials{
-				&mockInsecureTLSCredentials{}, // TODO(Joerger) replace insecure credentials
-			},
-			DialOpts: []grpc.DialOption{
-				grpc.WithTransportCredentials(insecure.NewCredentials()), // TODO(Joerger) remove insecure dial option
-			},
+		modifyConfig: func(c *Config) {
+			c.Addrs = nil
 		},
 		assertErr: func(t require.TestingT, err error, _ ...interface{}) {
 			require.Error(t, err)
-			require.Contains(t, err.Error(), "no connection methods found, try providing Dialer or Addrs in config")
+			require.ErrorContains(t, err, "no connection methods found, try providing Dialer or Addrs in config")
 		},
 	}}
 
 	for _, tt := range tests {
 		t.Run(tt.desc, func(t *testing.T) {
-			clt, err := srv.NewClient(ctx, WithConfig(tt.config))
-			tt.assertErr(t, err)
+			cfg := srv.clientCfg()
+			tt.modifyConfig(&cfg)
 
-			if err == nil {
-				t.Cleanup(func() { require.NoError(t, clt.Close()) })
-				// requests to the server should succeed.
-				_, err = clt.Ping(ctx)
-				require.NoError(t, err)
+			clt, err := New(ctx, cfg)
+			tt.assertErr(t, err)
+			if err != nil {
+				return
 			}
+
+			// Requests to the server should succeed.
+			_, err = clt.Ping(ctx)
+			assert.NoError(t, err, "Ping failed")
+			assert.NoError(t, clt.Close(), "Close failed")
 		})
 	}
 }
@@ -123,22 +116,16 @@ func TestNewDialBackground(t *testing.T) {
 	t.Parallel()
 	ctx := context.Background()
 
-	// get listener but don't serve it yet.
+	// Create a server but don't serve it yet.
 	l, err := net.Listen("tcp", "localhost:")
 	require.NoError(t, err)
 	addr := l.Addr().String()
+	srv := newMockServer(t, addr, &pingService{})
 
 	// Create client before the server is listening.
-	clt, err := New(ctx, Config{
-		DialInBackground: true,
-		Addrs:            []string{addr},
-		Credentials: []Credentials{
-			&mockInsecureTLSCredentials{}, // TODO(Joerger) replace insecure credentials
-		},
-		DialOpts: []grpc.DialOption{
-			grpc.WithTransportCredentials(insecure.NewCredentials()), // TODO(Joerger) remove insecure dial option
-		},
-	})
+	cfg := srv.clientCfg()
+	cfg.DialInBackground = true
+	clt, err := New(ctx, cfg)
 	require.NoError(t, err)
 	t.Cleanup(func() { require.NoError(t, clt.Close()) })
 
@@ -148,8 +135,8 @@ func TestNewDialBackground(t *testing.T) {
 	_, err = clt.Ping(cancelCtx)
 	require.Error(t, err)
 
-	// Start the server and wait for the client connection to be ready.
-	startMockServerWithListener(t, l)
+	// Server the listener and wait for the client connection to be ready.
+	srv.serve(t, l)
 	require.NoError(t, clt.waitForConnectionReady(ctx))
 
 	// requests to the server should succeed.
@@ -161,32 +148,27 @@ func TestWaitForConnectionReady(t *testing.T) {
 	t.Parallel()
 	ctx := context.Background()
 
+	// Create a server but don't serve it yet.
 	l, err := net.Listen("tcp", "localhost:")
 	require.NoError(t, err)
 	addr := l.Addr().String()
+	srv := newMockServer(t, addr, &proto.UnimplementedAuthServiceServer{})
 
 	// Create client before the server is listening.
-	clt, err := New(ctx, Config{
-		DialInBackground: true,
-		Addrs:            []string{addr},
-		Credentials: []Credentials{
-			&mockInsecureTLSCredentials{}, // TODO(Joerger) replace insecure credentials
-		},
-		DialOpts: []grpc.DialOption{
-			grpc.WithTransportCredentials(insecure.NewCredentials()), // TODO(Joerger) remove insecure dial option
-		},
-	})
+	cfg := srv.clientCfg()
+	cfg.DialInBackground = true
+	clt, err := New(ctx, cfg)
 	require.NoError(t, err)
 	t.Cleanup(func() { require.NoError(t, clt.Close()) })
 
-	// WaitForConnectionReady should return false once the
+	// WaitForConnectionReady should return an error once the
 	// context is canceled if the server isn't open to connections.
-	cancelCtx, cancel := context.WithTimeout(ctx, time.Second*3)
+	cancelCtx, cancel := context.WithTimeout(ctx, 100*time.Millisecond)
 	defer cancel()
 	require.Error(t, clt.waitForConnectionReady(cancelCtx))
 
 	// WaitForConnectionReady should return nil if the server is open to connections.
-	startMockServerWithListener(t, l)
+	srv.serve(t, l)
 	require.NoError(t, clt.waitForConnectionReady(ctx))
 
 	// WaitForConnectionReady should return an error if the grpc connection is closed.
@@ -194,10 +176,282 @@ func TestWaitForConnectionReady(t *testing.T) {
 	require.Error(t, clt.waitForConnectionReady(ctx))
 }
 
+type listResourcesService struct {
+	*proto.UnimplementedAuthServiceServer
+}
+
+func (s *listResourcesService) ListResources(ctx context.Context, req *proto.ListResourcesRequest) (*proto.ListResourcesResponse, error) {
+	resources, err := testResources[types.ResourceWithLabels](req.ResourceType, req.Namespace)
+	if err != nil {
+		return nil, trail.ToGRPC(err)
+	}
+
+	resp := &proto.ListResourcesResponse{
+		Resources:  make([]*proto.PaginatedResource, 0, len(resources)),
+		TotalCount: int32(len(resources)),
+	}
+
+	var (
+		takeResources    = req.StartKey == ""
+		lastResourceName string
+	)
+	for _, resource := range resources {
+		if resource.GetName() == req.StartKey {
+			takeResources = true
+			continue
+		}
+
+		if !takeResources {
+			continue
+		}
+
+		var protoResource *proto.PaginatedResource
+		switch req.ResourceType {
+		case types.KindDatabaseServer:
+			database, ok := resource.(*types.DatabaseServerV3)
+			if !ok {
+				return nil, trace.Errorf("database server has invalid type %T", resource)
+			}
+
+			protoResource = &proto.PaginatedResource{Resource: &proto.PaginatedResource_DatabaseServer{DatabaseServer: database}}
+		case types.KindAppServer:
+			app, ok := resource.(*types.AppServerV3)
+			if !ok {
+				return nil, trace.Errorf("application server has invalid type %T", resource)
+			}
+
+			protoResource = &proto.PaginatedResource{Resource: &proto.PaginatedResource_AppServer{AppServer: app}}
+		case types.KindNode:
+			srv, ok := resource.(*types.ServerV2)
+			if !ok {
+				return nil, trace.Errorf("node has invalid type %T", resource)
+			}
+
+			protoResource = &proto.PaginatedResource{Resource: &proto.PaginatedResource_Node{Node: srv}}
+		case types.KindKubeServer:
+			srv, ok := resource.(*types.KubernetesServerV3)
+			if !ok {
+				return nil, trace.Errorf("kubernetes server has invalid type %T", resource)
+			}
+
+			protoResource = &proto.PaginatedResource{Resource: &proto.PaginatedResource_KubernetesServer{KubernetesServer: srv}}
+		case types.KindWindowsDesktop:
+			desktop, ok := resource.(*types.WindowsDesktopV3)
+			if !ok {
+				return nil, trace.Errorf("windows desktop has invalid type %T", resource)
+			}
+
+			protoResource = &proto.PaginatedResource{Resource: &proto.PaginatedResource_WindowsDesktop{WindowsDesktop: desktop}}
+		case types.KindAppOrSAMLIdPServiceProvider:
+			appServerOrSP, ok := resource.(*types.AppServerOrSAMLIdPServiceProviderV1)
+			if !ok {
+				return nil, trace.Errorf("AppServerOrSAMLIdPServiceProvider has invalid type %T", resource)
+			}
+
+			protoResource = &proto.PaginatedResource{Resource: &proto.PaginatedResource_AppServerOrSAMLIdPServiceProvider{AppServerOrSAMLIdPServiceProvider: appServerOrSP}}
+		}
+		resp.Resources = append(resp.Resources, protoResource)
+		lastResourceName = resource.GetName()
+		if len(resp.Resources) == int(req.Limit) {
+			break
+		}
+	}
+
+	if len(resp.Resources) != len(resources) {
+		resp.NextKey = lastResourceName
+	}
+
+	return resp, nil
+}
+
+const fiveMBNode = "fiveMBNode"
+
+func testResources[T types.ResourceWithLabels](resourceType, namespace string) ([]T, error) {
+	size := 50
+	// Artificially make each node ~ 100KB to force
+	// ListResources to fail with chunks of >= 40.
+	labelSize := 100000
+	resources := make([]T, 0, size)
+
+	switch resourceType {
+	case types.KindDatabaseServer:
+		for i := 0; i < size; i++ {
+			resource, err := types.NewDatabaseServerV3(types.Metadata{
+				Name: fmt.Sprintf("db-%d", i),
+				Labels: map[string]string{
+					"label": string(make([]byte, labelSize)),
+				},
+			}, types.DatabaseServerSpecV3{
+				Hostname: "localhost",
+				HostID:   fmt.Sprintf("host-%d", i),
+				Database: &types.DatabaseV3{
+					Metadata: types.Metadata{
+						Name: fmt.Sprintf("db-%d", i),
+					},
+					Spec: types.DatabaseSpecV3{
+						Protocol: types.DatabaseProtocolPostgreSQL,
+						URI:      "localhost",
+					},
+				},
+			})
+			if err != nil {
+				return nil, trace.Wrap(err)
+			}
+
+			resources = append(resources, any(resource).(T))
+		}
+	case types.KindAppServer:
+		for i := 0; i < size; i++ {
+			app, err := types.NewAppV3(types.Metadata{
+				Name: fmt.Sprintf("app-%d", i),
+			}, types.AppSpecV3{
+				URI: "localhost",
+			})
+			if err != nil {
+				return nil, trace.Wrap(err)
+			}
+
+			resource, err := types.NewAppServerV3(types.Metadata{
+				Name: fmt.Sprintf("app-%d", i),
+				Labels: map[string]string{
+					"label": string(make([]byte, labelSize)),
+				},
+			}, types.AppServerSpecV3{
+				HostID: fmt.Sprintf("host-%d", i),
+				App:    app,
+			})
+			if err != nil {
+				return nil, trace.Wrap(err)
+			}
+
+			resources = append(resources, any(resource).(T))
+		}
+	case types.KindNode:
+		for i := 0; i < size; i++ {
+			nodeLabelSize := labelSize
+			if namespace == fiveMBNode && i == 0 {
+				// Artificially make a node ~ 5MB to force
+				// ListNodes to fail regardless of chunk size.
+				nodeLabelSize = 5000000
+			}
+
+			var err error
+			resource, err := types.NewServerWithLabels(fmt.Sprintf("node-%d", i), types.KindNode, types.ServerSpecV2{},
+				map[string]string{
+					"label": string(make([]byte, nodeLabelSize)),
+				},
+			)
+			if err != nil {
+				return nil, trace.Wrap(err)
+			}
+
+			resources = append(resources, any(resource).(T))
+		}
+	case types.KindKubeServer:
+		for i := 0; i < size; i++ {
+			var err error
+			name := fmt.Sprintf("kube-service-%d", i)
+			kube, err := types.NewKubernetesClusterV3(types.Metadata{
+				Name:   name,
+				Labels: map[string]string{"name": name},
+			},
+				types.KubernetesClusterSpecV3{},
+			)
+			if err != nil {
+				return nil, trace.Wrap(err)
+			}
+			resource, err := types.NewKubernetesServerV3(
+				types.Metadata{
+					Name: name,
+					Labels: map[string]string{
+						"label": string(make([]byte, labelSize)),
+					},
+				},
+				types.KubernetesServerSpecV3{
+					HostID:  fmt.Sprintf("host-%d", i),
+					Cluster: kube,
+				},
+			)
+			if err != nil {
+				return nil, trace.Wrap(err)
+			}
+
+			resources = append(resources, any(resource).(T))
+		}
+	case types.KindWindowsDesktop:
+		for i := 0; i < size; i++ {
+			var err error
+			name := fmt.Sprintf("windows-desktop-%d", i)
+			resource, err := types.NewWindowsDesktopV3(
+				name,
+				map[string]string{"label": string(make([]byte, labelSize))},
+				types.WindowsDesktopSpecV3{
+					Addr:   "_",
+					HostID: "_",
+				})
+			if err != nil {
+				return nil, trace.Wrap(err)
+			}
+
+			resources = append(resources, any(resource).(T))
+		}
+	case types.KindAppOrSAMLIdPServiceProvider:
+		for i := 0; i < size; i++ {
+			// Alternate between adding Apps and SAMLIdPServiceProviders. If `i` is even, add an app.
+			if i%2 == 0 {
+				app, err := types.NewAppV3(types.Metadata{
+					Name: fmt.Sprintf("app-%d", i),
+				}, types.AppSpecV3{
+					URI: "localhost",
+				})
+				if err != nil {
+					return nil, trace.Wrap(err)
+				}
+
+				appServer, err := types.NewAppServerV3(types.Metadata{
+					Name: fmt.Sprintf("app-%d", i),
+					Labels: map[string]string{
+						"label": string(make([]byte, labelSize)),
+					},
+				}, types.AppServerSpecV3{
+					HostID: fmt.Sprintf("host-%d", i),
+					App:    app,
+				})
+				if err != nil {
+					return nil, trace.Wrap(err)
+				}
+
+				resource := &types.AppServerOrSAMLIdPServiceProviderV1{
+					Resource: &types.AppServerOrSAMLIdPServiceProviderV1_AppServer{
+						AppServer: appServer,
+					},
+				}
+
+				resources = append(resources, any(resource).(T))
+			} else {
+				sp := &types.SAMLIdPServiceProviderV1{ResourceHeader: types.ResourceHeader{Metadata: types.Metadata{Name: fmt.Sprintf("saml-app-%d", i), Labels: map[string]string{
+					"label": string(make([]byte, labelSize)),
+				}}}}
+
+				resource := &types.AppServerOrSAMLIdPServiceProviderV1{
+					Resource: &types.AppServerOrSAMLIdPServiceProviderV1_SAMLIdPServiceProvider{
+						SAMLIdPServiceProvider: sp,
+					},
+				}
+				resources = append(resources, any(resource).(T))
+			}
+		}
+	default:
+		return nil, trace.Errorf("unsupported resource type %s", resourceType)
+	}
+
+	return resources, nil
+}
+
 func TestListResources(t *testing.T) {
 	t.Parallel()
 	ctx := context.Background()
-	srv := startMockServer(t)
+	srv := startMockServer(t, &listResourcesService{})
 
 	testCases := map[string]struct {
 		resourceType   string
@@ -226,7 +480,7 @@ func TestListResources(t *testing.T) {
 	}
 
 	// Create client
-	clt, err := srv.NewClient(ctx)
+	clt, err := New(ctx, srv.clientCfg())
 	require.NoError(t, err)
 
 	for name, test := range testCases {
@@ -298,10 +552,11 @@ func testGetResources[T types.ResourceWithLabels](t *testing.T, clt *Client, kin
 
 func TestGetResources(t *testing.T) {
 	t.Parallel()
-	srv := startMockServer(t)
+	ctx := context.Background()
+	srv := startMockServer(t, &listResourcesService{})
 
 	// Create client
-	clt, err := srv.NewClient(context.Background())
+	clt, err := New(ctx, srv.clientCfg())
 	require.NoError(t, err)
 
 	t.Run("DatabaseServer", func(t *testing.T) {
@@ -338,10 +593,10 @@ func TestGetResources(t *testing.T) {
 func TestGetResourcesWithFilters(t *testing.T) {
 	t.Parallel()
 	ctx := context.Background()
-	srv := startMockServer(t)
+	srv := startMockServer(t, &listResourcesService{})
 
 	// Create client
-	clt, err := srv.NewClient(ctx)
+	clt, err := New(ctx, srv.clientCfg())
 	require.NoError(t, err)
 
 	testCases := map[string]struct {
@@ -393,12 +648,4 @@ func TestGetResourcesWithFilters(t *testing.T) {
 			require.Empty(t, cmp.Diff(expectedResources, resources))
 		})
 	}
-}
-
-func TestMain(m *testing.M) {
-	flag.Parse()
-	if testing.Verbose() {
-		logrus.SetLevel(logrus.DebugLevel)
-	}
-	os.Exit(m.Run())
 }
